Fix params hashing (#2540)
* move params freezing to separated module

* freeze dict for toml parser

* FIX FrozenOrderedDict name

* Freeze only params of config, not all config

* +docstring
