+++ luigi/configuration/toml_parser.py
@@ -22,6 +22,7 @@ except ImportError:
     toml = False
 
 from .base_parser import BaseParser
+from ..freezing import recursively_freeze
 
 
 class LuigiTomlParser(BaseParser):
@@ -51,6 +52,13 @@ class LuigiTomlParser(BaseParser):
         for path in config_paths:
             if os.path.isfile(path):
                 self.data = self._update_data(self.data, toml.load(path))
+
+        # freeze dict params
+        for section, content in self.data.items():
+            for key, value in content.items():
+                if isinstance(value, dict):
+                    self.data[section][key] = recursively_freeze(value)
+
         return self.data
 
     def get(self, section, option, default=NO_DEFAULT, **kwargs):
