+++ test/testconfig/luigi.toml
@@ -5,3 +5,6 @@ logging_conf_file = "test/testconfig/logging.cfg"
 client = "hadoopcli"
 snakebite_autoconfig = false
 namenode_host = "must be overridden in local config"
+
+[SomeTask]
+param = {key1 = "value1", key2 = "value2"}
